refactor: align resource processor module names
