+++ reme/steps/evolve/__init__.py
@@ -1,11 +1,11 @@
 """Evolve steps."""
 
 from ._evolve import now
-from .auto_image import AutoImageResourceStep
+from .auto_image_resource import AutoImageResourceStep
 from .auto_memory import AutoMemoryStep
 from .auto_memory_cc import AutoMemoryCCStep
 from .auto_resource import AutoResourceStep
-from .auto_text import AutoTextResourceStep
+from .auto_text_resource import AutoTextResourceStep
 from .compressor import CompressorStep
 from .dream import DreamExtractStep, DreamFinishStep, DreamIntegrateStep, DreamTopicsStep, ProactiveStep
