+++ reme/steps/evolve/auto_image_resource.yaml
