+++ tests/unit/test_background_steps.py
@@ -33,7 +33,7 @@ from reme.enumeration import ComponentEnum
 from reme.steps.evolve.auto_memory import AutoMemoryStep
 from reme.steps.evolve._auto_resource import _compute_note_stem
 from reme.steps.evolve.auto_resource import AutoResourceStep
-from reme.steps.evolve.auto_text import AutoTextResourceStep
+from reme.steps.evolve.auto_text_resource import AutoTextResourceStep
 from reme.steps.file_io.daily_list import DailyListStep
 from reme.steps.file_io.frontmatter_update import FrontmatterUpdateStep
 from reme.steps.file_io.move import MoveStep
